requirements.txt
flake8==3.3.0
lxml==3.8.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.5.0
PyYAML==3.12
six==1.10.0
webencodings==0.5.1
xmltodict==0.11.0
